Zip File of Schematic and PCB
This zip file includes the Ki-CAD files associated with the PCB I designed for the portable phone charger. Each section of the schematic is compartmentalized and labeled along with the layout containing a double-sided PCB. The front side of the PCB has the micro USB, charging IC, protection IC and battery pads. The back of the PCB contains the push button switch, boost converter, and LED driver.